AI_VIDEO_BUSINESS/clinic-content-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinic-content-engine"
version = "0.1.0"
description = "Compliant AI-powered content engine for clinics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "openai>=1.40.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0"
]

[project.scripts]
cce = "cce.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
